Move extractVariables utility
