+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/expression/ConnectorExpressions.java
@@ -14,20 +14,35 @@
 package io.trino.plugin.base.expression;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.graph.SuccessorsFunction;
+import com.google.common.graph.Traverser;
 import io.trino.spi.expression.Call;
 import io.trino.spi.expression.ConnectorExpression;
+import io.trino.spi.expression.Variable;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Stream;
 
+import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.getOnlyElement;
+import static com.google.common.collect.Streams.stream;
 import static io.trino.spi.expression.Constant.TRUE;
 import static io.trino.spi.expression.StandardFunctions.AND_FUNCTION_NAME;
 import static io.trino.spi.type.BooleanType.BOOLEAN;
+import static java.util.Objects.requireNonNull;
 
 public final class ConnectorExpressions
 {
     private ConnectorExpressions() {}
+
+    public static List<Variable> extractVariables(ConnectorExpression expression)
+    {
+        return preOrder(expression)
+                .filter(Variable.class::isInstance)
+                .map(Variable.class::cast)
+                .collect(toImmutableList());
+    }
 
     public static List<ConnectorExpression> extractConjuncts(ConnectorExpression expression)
     {
@@ -63,5 +78,12 @@ public final class ConnectorExpressions
             return TRUE;
         }
         return getOnlyElement(expressions);
+    }
+
+    private static Stream<ConnectorExpression> preOrder(ConnectorExpression expression)
+    {
+        return stream(
+                Traverser.forTree((SuccessorsFunction<ConnectorExpression>) ConnectorExpression::getChildren)
+                        .depthFirstPreOrder(requireNonNull(expression, "expression is null")));
     }
 }
